[chore]: addmenu/component의 파일들 패키징
