+++ app/src/main/java/com/kuit/ourmenu/ui/addmenu/component/item/AddMenuInfoTextFieldItem.kt
@@ -1,25 +1,26 @@
-package com.kuit.ourmenu.ui.addmenu.component
+package com.kuit.ourmenu.ui.addmenu.component.item
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.kuit.ourmenu.R
@@ -32,9 +33,12 @@ import com.kuit.ourmenu.ui.theme.Primary500Main
 import com.kuit.ourmenu.ui.theme.ourMenuTypography
 
 @Composable
-fun AddMenuInfoMenuBoardFieldItem(
+fun AddMenuInfoTextFieldItem(
+    fieldName: String,
     text: String,
     onTextChange: (String) -> Unit,
+    placeholder: String,
+    isPriceInfo: Boolean = false,
 ) {
     Column(
         modifier = Modifier
@@ -48,7 +52,7 @@ fun AddMenuInfoMenuBoardFieldItem(
                 .padding(bottom = 4.dp)
         ) {
             Text(
-                text = stringResource(R.string.menuboard),
+                text = fieldName,
                 style = ourMenuTypography().pretendard_600_14
             )
             Text(
@@ -57,7 +61,6 @@ fun AddMenuInfoMenuBoardFieldItem(
                 color = Primary500Main
             )
         }
-        //드롭다운으로 수정할 예정
         CustomTextField(
             modifier = Modifier
                 .fillMaxWidth()
@@ -70,18 +73,18 @@ fun AddMenuInfoMenuBoardFieldItem(
             containerColor = Neutral100,
             placeHolder = {
                 Text(
-                    text = stringResource(R.string.save_menuboard),
+                    text = placeholder,
                     style = ourMenuTypography().pretendard_500_14,
                     color = Neutral500
                 )
             },
             textStyle = ourMenuTypography().pretendard_700_14.copy(color = Neutral700),
-            trailingIcon = {
-                Icon(
-                    painter = painterResource(R.drawable.ic_searchbar_search),
-                    contentDescription = "search icon",
-                    tint = Color.Unspecified
-                )
+            keyboardOptions = if(isPriceInfo) {
+                KeyboardOptions(
+                    keyboardType = KeyboardType.Number,
+                    )
+            }else{
+                KeyboardOptions.Default
             }
         )
 
@@ -90,10 +93,28 @@ fun AddMenuInfoMenuBoardFieldItem(
 
 @Preview(showBackground = true)
 @Composable
-private fun AddMenuInfoMenuBoardFieldItemPreview() {
-    var text by rememberSaveable { mutableStateOf("") }
-    AddMenuInfoMenuBoardFieldItem(
-        text = text,
-        onTextChange = { text = it }
-    )
+private fun AddMenuInfoTextFieldItemPreview() {
+    var priceText by rememberSaveable { mutableStateOf("") }
+    var restaurantNameText by rememberSaveable { mutableStateOf("") }
+
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        AddMenuInfoTextFieldItem(
+            fieldName = stringResource(R.string.menu_price),
+            text = priceText,
+            onTextChange = { priceText = it },
+            placeholder = stringResource(R.string.type_menu_price),
+            isPriceInfo = true
+        )
+        AddMenuInfoTextFieldItem(
+            fieldName = stringResource(R.string.restaurant_name),
+            text = restaurantNameText,
+            onTextChange = { restaurantNameText = it },
+            placeholder = stringResource(R.string.type_restaurant_name)
+        )
+    }
+
 }
