+++ app/src/main/java/com/kuit/ourmenu/ui/addmenu/screen/AddMenuTagScreen.kt
@@ -41,7 +41,7 @@ import androidx.compose.ui.unit.dp
 import com.kuit.ourmenu.R
 import com.kuit.ourmenu.ui.addmenu.component.AddMenuTopAppBar
 import com.kuit.ourmenu.ui.addmenu.component.IconSelectBottomSheet
-import com.kuit.ourmenu.ui.addmenu.component.TagSelectBottomSheet
+import com.kuit.ourmenu.ui.addmenu.component.bottomsheet.TagSelectBottomSheet
 import com.kuit.ourmenu.ui.common.CustomTextField
 import com.kuit.ourmenu.ui.theme.Neutral100
 import com.kuit.ourmenu.ui.theme.Neutral300
